Factore:Ticket055 - backend update e delete com restrição para user

+++ backend/src/services/comments/delete_comments.service.ts
@@ -4,18 +4,32 @@ import { Comments } from "../../entities/comments.entity";
 import { AppError } from "../../errors/appError";
 import { IAdverts, IAdvertsDelete } from "../../interfaces/adverts";
 
-const deleteCommentsService = async (id: string ): Promise<string> => {
+const deleteCommentsService = async (id: string, userId:string ): Promise<string> => {
   const commentsRepository = AppDataSource.getRepository(Comments);
 
-  const deletedComments = await commentsRepository.findOneBy({ id });
-
+  const deletedComments = await commentsRepository.findOne({
+    where: {id:id
+    } ,
+    relations:{
+      user: true
+    }
+    });
+  
   if (!deletedComments) {
     throw new AppError("Comment not found");
   }
-
+  
+  
+  if( deletedComments.user.id !== userId){
+    throw new AppError("User  Unauthorized");
+  }
+    
   await commentsRepository.delete({ id });
+    
+    return "Comments deleted";
 
-  return "Comments deleted";
+
+
 };
 
 export default deleteCommentsService;
